+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 2.5.0 - 2021-11-08
+## Unreleased
 
 ### Added
 - New dropdown to limit the number of languages that are shown side by side in the overview